requirements.txt
beautifulsoup4==4.7.0
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
Django==1.11.18
filebrowser_safe==0.5.0
future==0.17.1
grappelli_safe==0.5.1
idna==2.8
Mezzanine==4.3.1
oauthlib==2.1.0
Pillow==5.4.0
requests==2.21.0
requests_oauthlib==1.0.0
solrpy=0.9.9
soupsieve==1.6.2
urllib3==1.24.1
pytz==2018.7
tzlocal==1.5.1
webencodings==0.5.1
